LayoutTests/storage/indexeddb/intversion-and-setversion-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test that int version upgrades a string version and that string versions are then disallowed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-and-setversion.html"
indexedDB.deleteDatabase(dbname)
FAIL String(request) should be [object IDBOpenDBRequest]. Was [object IDBVersionChangeRequest].

Call open with no version parameter:
indexedDB.open(dbname)

initialUpgradeNeeded():

openSuccess():
db = event.target.result
PASS db.version is 1
request = db.setVersion("some version")

inSetVersion():
PASS db.version is "some version"

setVersionDone():
PASS db.version is "some version"
db.close()
indexedDB.open(dbname)
request.onsuccess = secondOpen

secondOpen():
db = event.target.result
PASS String(db) is "[object IDBDatabase]"
PASS db.version is "some version"
db.close()
request = indexedDB.open(dbname, 2)

firstUpgradeNeeded():
db = event.target.result
PASS String(db) is "[object IDBDatabase]"
PASS String(request.transaction) is "[object IDBTransaction]"
FAIL event.oldVersion should be some version (of type string). Was 0 (of type number).
PASS event.newVersion is 2
request.transaction.oncomplete = transactionCompleted

transactionCompleted():
didTransactionComplete = true

firstOpenWithVersion():
PASS didTransactionComplete is true
db = event.target.result
db.onversionchange = versionChangeGoingFromStringToInt
PASS String(db) is "[object IDBDatabase]"
PASS request.transaction is null
PASS db.version is 2
PASS String(request) is "[object IDBOpenDBRequest]"
request = db.setVersion("string version 2")
PASS String(request) is "[object IDBVersionChangeRequest]"
request.onsuccess = secondSetVersionCallback

secondSetVersionCallback():
transaction = event.target.result
transaction.oncomplete = setIntVersion2

setIntVersion2():
request = indexedDB.open(dbname, 2)
request.onsuccess = version2ConnectionSuccess
request.onblocked = version2ConnectionBlocked

versionChangeGoingFromStringToInt():
db.close()

version2ConnectionBlocked():

version2ConnectionSuccess():
event.target.result.close()
request = indexedDB.open(dbname, 1)
request.onerror = errorWhenTryingLowVersion

errorWhenTryingLowVersion():
request.webkitErrorMessage = The requested version (1) is less than the existing version (2).
request = indexedDB.open(dbname, 4)
request.onupgradeneeded = secondUpgradeNeeded
request.onsuccess = thirdOpenSuccess

secondUpgradeNeeded():
gotSecondUpgradeNeeded = true
PASS event.oldVersion is 2
PASS event.newVersion is 4

thirdOpenSuccess():
PASS gotSecondUpgradeNeeded is true
db = event.target.result
PASS db.version is 4
db.close()
request = indexedDB.open(dbname)
request.onsuccess = fourthOpenSuccess

fourthOpenSuccess():
db = event.target.result
PASS db.version is 4
PASS successfullyParsed is true

TEST COMPLETE
